Extract interface. Make classes package private.
